One more update related to the #5381
